requirements.txt
# openai
openai
tiktoken

# langchain
langchain
docx2txt
unstructured
pypdf
pdf2image
pdfminer
pdfminer.six

# huggingface
# huggingface-hub
# tokenizers
# sentence-transformers

# evaluate
evaluate
Cython
rouge_score
absl-py

# misc
pydantic
pandas
numpy
scikit-learn
python-dotenv
chromadb

# agents
# wikipedia
# filetype
# llama-index
# faiss-cpu